[REFACTOR] Location Modal Refactor Done
